archive/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reduck"
version = "0.1.0"
description = "Voice interface for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "pydantic>=2.0",
    "claude-agent-sdk",
    "python-dotenv>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["reduck"]

[project.scripts]
reduck = "reduck.cli:main"
